% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chop.R
\name{chop}
\alias{chop}
\title{Split polygon into many}
\usage{
chop(x, cellsize = 10, cells.dim = 20)
}
\arguments{
\item{x}{Spatial object}

\item{cellsize}{size}

\item{cells.dim}{Dimension}
}
\description{
Split polygon into many
}
\details{
Works on spatial classes of type \code{SpatialPolygons}, and on
Well-Known Text character strings.
}
\examples{
library("rgeos")
wkt <- "POLYGON((-180 -20, -140 55, 10 0, -140 -60, -180 -20))"

# SpatialPolygons input
poly <- readWKT(wkt)
plot(poly)
polys <- chop(x = poly)
to_wkt(polys)
to_wkt(polys)[[2]]
plot(polys)
plot(poly, add = TRUE, lwd = 6)

# WKT character input
chop(wkt)

# geojson character input
# geojsonio::as.json(wellknown::wkt2geojson(wkt)$geometry)
x <- '{"type":"Polygon","coordinates":[[["-180.0","-20.0"],["-140.0","55.0"],["10.0","0.0"],["-140.0","-60.0"],["-180.0","-20.0"]]]}'
# chop(x)

# geojson list input
# chop(wkt)

## ignore
# geojsonio::geojson_json(poly) \%>\% geojsonio::map_leaf()
# geojsonio::geojson_json(polys) \%>\% geojsonio::map_leaf()
}

