% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_polygon.R
\name{split_polygon}
\alias{split_polygon}
\title{Split polygon}
\usage{
split_polygon(x, cellsize = 10, cells.dim = 20)
}
\arguments{
\item{x}{Spatial object}

\item{cellsize}{size}

\item{cells.dim}{Dimension}
}
\description{
Split polygon
}
\examples{
library("rgeos")
poly <- readWKT("POLYGON((-180 -20, -140 55, 10 0, -140 -60, -180 -20))")
plot(poly)
polys <- split_polygon(x = poly)
to_wkt(polys)
plot(readWKT(to_wkt(polys)[1]))
plot(readWKT(to_wkt(polys)[10]))
}

