% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_wkt.R
\name{to_wkt}
\alias{to_wkt}
\title{To WKT}
\usage{
to_wkt(x)
}
\arguments{
\item{x}{Input}
}
\description{
To WKT
}
\examples{
library("rgeos")
wkt <- "POLYGON((-180 -20, -140 55, 10 0, -140 -60, -180 -20))"
poly <- readWKT(wkt)
polys <- chop(x = poly)
to_wkt(polys)
to_wkt(polys)[[2]]
}

